backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.1
pymongo==4.6.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
email-validator==2.1.1
PyJWT==2.12.1
bcrypt==4.1.2
firebase-admin==6.5.0
